Create StatsDataModel's from CSV files via YAMTL

+++ yamtl.cafev.stats/src/main/emf-gen/dataDescription/NumericalType.java
@@ -28,12 +28,12 @@ public interface NumericalType extends StatsDataType {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Min</em>' attribute.
-	 * @see #setMin(float)
+	 * @see #setMin(double)
 	 * @see dataDescription.DataDescriptionPackage#getNumericalType_Min()
 	 * @model required="true"
 	 * @generated
 	 */
-	float getMin();
+	double getMin();
 
 	/**
 	 * Sets the value of the '{@link dataDescription.NumericalType#getMin <em>Min</em>}' attribute.
@@ -43,19 +43,19 @@ public interface NumericalType extends StatsDataType {
 	 * @see #getMin()
 	 * @generated
 	 */
-	void setMin(float value);
+	void setMin(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Max</b></em>' attribute.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Max</em>' attribute.
-	 * @see #setMax(float)
+	 * @see #setMax(double)
 	 * @see dataDescription.DataDescriptionPackage#getNumericalType_Max()
 	 * @model required="true"
 	 * @generated
 	 */
-	float getMax();
+	double getMax();
 
 	/**
 	 * Sets the value of the '{@link dataDescription.NumericalType#getMax <em>Max</em>}' attribute.
@@ -65,19 +65,19 @@ public interface NumericalType extends StatsDataType {
 	 * @see #getMax()
 	 * @generated
 	 */
-	void setMax(float value);
+	void setMax(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Mean</b></em>' attribute.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Mean</em>' attribute.
-	 * @see #setMean(float)
+	 * @see #setMean(double)
 	 * @see dataDescription.DataDescriptionPackage#getNumericalType_Mean()
 	 * @model required="true"
 	 * @generated
 	 */
-	float getMean();
+	double getMean();
 
 	/**
 	 * Sets the value of the '{@link dataDescription.NumericalType#getMean <em>Mean</em>}' attribute.
@@ -87,19 +87,19 @@ public interface NumericalType extends StatsDataType {
 	 * @see #getMean()
 	 * @generated
 	 */
-	void setMean(float value);
+	void setMean(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Std</b></em>' attribute.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Std</em>' attribute.
-	 * @see #setStd(float)
+	 * @see #setStd(double)
 	 * @see dataDescription.DataDescriptionPackage#getNumericalType_Std()
 	 * @model required="true"
 	 * @generated
 	 */
-	float getStd();
+	double getStd();
 
 	/**
 	 * Sets the value of the '{@link dataDescription.NumericalType#getStd <em>Std</em>}' attribute.
@@ -109,6 +109,6 @@ public interface NumericalType extends StatsDataType {
 	 * @see #getStd()
 	 * @generated
 	 */
-	void setStd(float value);
+	void setStd(double value);
 
 } // NumericalType
